+++ NEWS.md
@@ -1,5 +1,7 @@
 # ussie 0.0.0.9000
 
+* Added functions, `uss_points()` and `uss_points_per_win()`. 
+
 * Added `uss_make_teams_matches()` to convert a `matches` tibble to a `teams-matches` tibble. (#1)
 
 * Changed `_games` to `_matches`, e.g. `uss_make_matches()`. (#2)